requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.7
boto3==1.28.57
Werkzeug==2.3.7
gunicorn==21.2.0